Address android-reviewer findings (round 2)
- Fix AcwMapOutputFile: add property to GenerateTrimmableTypeMap task,
  wire through TrimmableTypeMapGenerator.Execute, and write acw-map.txt
  via AcwMapWriter so _ConvertCustomView and _UpdateAndroidResgen get
  real managed→Java mappings
- Split ManifestAttributeInfo.cs: move each of the 8 types to its own
  file (one type per file rule)
- Make ManifestGenerator internal: no external consumers; tests use
  InternalsVisibleTo
- Return IReadOnlyList<string> from public Generate overloads (was IList)
