+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Scanner/PermissionTreeInfo.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+using System.Collections.Generic;
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap;
+
+public sealed record PermissionTreeInfo
+{
+	public required string Name { get; init; }
+	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
+}
