+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Scanner/PropertyInfo.cs
@@ -1,0 +1,10 @@
+#nullable enable
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap;
+
+public sealed record PropertyInfo
+{
+	public required string Name { get; init; }
+	public string? Value { get; init; }
+	public string? Resource { get; init; }
+}
