+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Scanner/UsesConfigurationInfo.cs
@@ -1,0 +1,12 @@
+#nullable enable
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap;
+
+public sealed record UsesConfigurationInfo
+{
+	public bool ReqFiveWayNav { get; init; }
+	public bool ReqHardKeyboard { get; init; }
+	public string? ReqKeyboardType { get; init; }
+	public string? ReqNavigation { get; init; }
+	public string? ReqTouchScreen { get; init; }
+}
