+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Scanner/UsesPermissionInfo.cs
@@ -1,0 +1,9 @@
+#nullable enable
+
+namespace Microsoft.Android.Sdk.TrimmableTypeMap;
+
+public sealed record UsesPermissionInfo
+{
+	public required string Name { get; init; }
+	public int? MaxSdkVersion { get; init; }
+}
